Split gmp/locale.js into distinct js modules
Split language and date parts into separated js modules.
